adding option to run PROSPECT-PRO depending on inputs
if CBC and PROT are both zero (default), then PROSPECT-5D is executed as
it was the case in the original version of SPART. Tests confirmed that
the result are still the same when running PROSPECT-5D now.

only if CBC and/or PROT are non-zero PROSPECT-PRO is executed. The dry
matter content Cdm must then be zero.

This way SPART now supports PROSPECT5-D (default) and PROSPECT-PRO (opt)
as leaf models.